move RecentlyLikedTileFeed to ui-connected
