+++ packages/apollos-ui-connected/src/RecentlyLikedTileFeed/RecentlyLikedTileFeed.js
@@ -16,7 +16,7 @@ import {
   withIsLoading,
 } from '@apollosproject/ui-kit';
 
-import HorizontalContentCardConnected from '../../../ui/HorizontalContentCardConnected';
+import HorizontalContentCardConnected from '../../../apolloschurchapp/src/ui/HorizontalContentCardConnected';
 
 const RowHeader = styled(({ theme }) => ({
   flexDirection: 'row',
