+++ packages/apollos-ui-connected/src/RecentlyLikedTileFeed/RecentlyLikedTileFeed.tests.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Providers from '../../../Providers';
+import { Providers } from '../utils/testUtils';
+
 import RecentlyLikedTileFeed from './RecentlyLikedTileFeed';
 
 describe('RecentlyLikedTileFeed', () => {
